git-svn-id: file:///Users/braun/svn/vermont/branches/vermont/lockfree@2547 aef3b71b-58ee-0310-9ba9-8811b9f0742f

+++ src/modules/ipfix/IpfixParser.cpp
@@ -715,7 +715,7 @@ int IpfixParser::processNetflowV9Packet(boost::shared_array<uint8_t> message, ui
 	}
 
 	msg(MSG_VDEBUG, "NetflowV9 message from %s contained %u Data Records and %u Template Records. Sequence number was %lu.", 
-		(sourceId->toString()).c_str(), numberOfDataRecords, numberOfTemplateRecords, sequenceNumber);
+		(sourceId->toString()).c_str(), numberOfDataRecords, numberOfTemplateRecords, (unsigned long) sequenceNumber);
 
 	// Update statistics
 	statTotalMessages++;
@@ -815,7 +815,7 @@ int IpfixParser::processIpfixPacket(boost::shared_array<uint8_t> message, uint16
 	}
 
 	msg(MSG_VDEBUG, "IPFIX message from %s contained %u Data Records and %u Template Records. Sequence number was %lu.", 
-		(sourceId->toString()).c_str(), numberOfDataRecords, numberOfTemplateRecords, sequenceNumber);
+		(sourceId->toString()).c_str(), numberOfDataRecords, numberOfTemplateRecords, (unsigned long) sequenceNumber);
 
 	// Update statistics
 	statTotalMessages++;
